Large Commit.....
This is the current base of the plugin in it's current state.  Unfortunately, current state is not working, as PCKPacker, and ZIPPacker does not actually build all of the files that are needed, to build the actual PCK or ZIP file, to exported, and have it work in the end base.  Therefore, we are changing how Exports are created.  Using export_presets.cfg, and Command Line execution of the Godot Executable to export the pack file.